ai/llm_runtime/requirements.txt
openai==0.28.0
pydantic==2.12.0
python-dotenv==1.1.1
